requirements.txt
streamlit>=1.32.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.38
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=0.21.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.0
